(see Dec) Added Entities
Currently, these entities don't serve any function; however, there
should soon be commits to change that. This commit merely serves to lay
the groundwork, and to keep the history easy to read.

The largest differences between this version and the text:
* Aircraft has an enum that is named EnemyType rather than Type.
* This enum contains one value for each of the different enemy sprites.
* There exists a shared_ptr to spriteMap, which is a mapping of each
enum of EnemyType to their respective row on the Sprite Sheet for easy
reference.
* Aircraft Sprites have their Scale and Origins hard-coded. All aircraft
in the game use Sprites of size 16x16, so there is no need for this to
vary.
(Entities, of course, are different.)
* The textbook seems to suggest that Entity either IS-A or HAS-A
SceneNode, but is somewhat unclear as to which, so this code implements
neither for now.
(I do believe the end result is inheritance, but I'm awaiting further
evidence.)
* SpriteNode is contained within SceneNode.h/cpp for convenience.
(Of course this is subject to change.)