(see Desc) Currently Debugging & Confused
Everything is built and hooked together. Using excessive cout
statements, I was able to determine that every function that ought to be
running, is running. Unfortunately, this does not impact the issue: our
new sprites are not being drawn. Why is that? It's not a missing
function call; running the code shows that the function that ought to
draw our sprites is being called. Additionally, our sprites are set up
in much the same way as our previous Sprites.

I have more digging to do, but for now:
* cout statements are littered everywhere. They're inside roughly every
function, and most of them are commented out at the moment.
* updateCurrent()'s single line of code is no longer commented out
* Our world is being updated within Galaga::update()
* All previous visuals are commented out of render()- never drawing them
to the screen- in case they are conflicting with what we are trying to
draw. Removing and re-adding them seems to have no impact, however.
* Switched around the order of function definitions in SceneNode.cpp
* Renamed draw() to drawCurrent() and drawCurrent() to draw() (the
textbook mistakenly swaps their names in the code)
* mWorldBounds are now hardcoded
* Our player aircraft is now moving at a velocity of 1,0 instead of
40,0.

Other oddities:
* It was observed that mWorldView.getSize() returns 0,0.  I've yet to
discover an explanation for this.