(see Desc) AHA!
window.getDefaultValue() is the root of all evil (in this case).
Apparently the default value for a view's size is (0,0) which is the
opposite of both what the online documentation and the textbook
describe. Meaning that as long as I create my own views, this won't be a
problem!
Additionally, views no longer update every "tick." If the view were
moving or changing, this would be necessary, but they are not!