Moved ResourceHolder.inl into ResourceHolder.h
Previously, I had the rather unfortunate, hacky fix of having the two
#include each other in order for the code to compile.
This worked well on my tower PC, but not my laptop... which was
interesting, considering they both were using the same compiler. Either
way, this was a clear indication to me that I needed a more stable fix.